deployed chart component seprated out
